http_uploader: complete curl init
Add constants for form setup, and finish form initialization code.

We should be good to call curl_easy_perform now; unfortunately this won't
build against the libcurl currently in-tree because we must upgrade to curl
v7.18.2 or higher in order to use CURLFORM_STREAM.  We must use
CURLFORM_STREAM to upload the file as part of a form-- to avoid this
we would have to send HTTP PUT instead of HTTP POST with form data.

Beyond that we need:
- some heuristic for distinguishing between catching up to the file's write
  pointer and file creation completion in FileReader::Read
- code for handling situations where FileReader::Read cannot return any
  data

Change-Id: Ib63b6431fc66e87f35cb27c7073ec20da527ccc7