webm_encoder_dshow: fix map indexes in CaptureSourceLoader
The map of capture sources filled by CaptureSourceLoader::FindAllSources
is intended to be used with 0 based consecutive indexes.  FindAllSources was
incrementing the index value used to assign capture device name when
skipping devices, which would potentially result in a map with names at, for
example, indexes 1, 3 and 6, with 0, 2, 4, and 5 empty.

Change-Id: Ib2693cecb9e8130b2a70f8da458b0d8d6d34b8af