webm_encoder_dshow: load capture device by name.
CaptureSourceLoader skips empty objects, which means direct use of the
index into the map of source devices when enumerating in
CaptureSourceLoader::GetSource was unreliable.  To reliably load the device,
its name must be used, so GetSource is now overloaded.
The version taking an integer argument passes the result of GetSourceName(index)
to the new version of GetSource, which uses a std::wstring, and finds a device a
matching the friendly name contained in the string.

Change-Id: I78fd3c286ca8336d734fc0d0a2347f1733006dea