Update patch set 9
Patch Set 9: Verified

Wow... that was a somewhat painful series of rebases. My fault-- WebmEncoderConfig::VideoCaptureConfig was needed in the AM_MEDIA_TYPE commit, and had to be moved back there from this commit, and then I dealt with comments+lint.  Should be ready for final review and landing.

Sorry for the gerrit spam if it sent a message for every patchset! :)

Patch-set: 9
Label: Code-Review=0
Label: Verified=+1